moved workflow panels to separate files
