+++ Localize/lang/strings.json
@@ -247,8 +247,8 @@
   "5fmV2Q": "Parameters used in Logic App will be converted into Azure Resource Manager template during deployment template generation.",
   "5gOG+F": "Provide the values to compare and select the operator to use.",
   "5hWg4V": "Select element",
-  "5lRHeK": "--",
   "5lEeZZ": "Channels",
+  "5lRHeK": "--",
   "5pSOjg": "Create workflows from template",
   "5qzZMo": "Loading...",
   "5rljf7": "Discovery Panel",
@@ -556,7 +556,6 @@
   "FLeN8N": "General",
   "FMA0Q0": "Your flow has been updated.",
   "FN5zHQ": "Returns a section of a string defined by the start index and the end index",
-  "FOoLuS": "Description",
   "FPBmrA": "Source schema node not found for node key ''{nodeKey}''",
   "FSD5Z4": "Select all workflows checkbox rows label",
   "FT3jt6": "Webhook reference information",
@@ -1446,8 +1445,8 @@
   "_5fmV2Q.comment": "Description for Workflow Parameters Part 2 for Legacy Parameters mode.",
   "_5gOG+F.comment": "Text description for how to use the Condition action.",
   "_5hWg4V.comment": "Select element",
-  "_5lRHeK.comment": "Accessibility label indicating that the value is not set",
   "_5lEeZZ.comment": "Channels tab title",
+  "_5lRHeK.comment": "Accessibility label indicating that the value is not set",
   "_5pSOjg.comment": "Panel header title for creating workflows",
   "_5qzZMo.comment": "A message shown at the bottom of a list when the next set of data is loading",
   "_5rljf7.comment": "Discovery Panel",
@@ -1755,7 +1754,6 @@
   "_FLeN8N.comment": "General section title",
   "_FMA0Q0.comment": "Chatbot report a bug button",
   "_FN5zHQ.comment": "Label for description of custom slice Function",
-  "_FOoLuS.comment": "Label for description",
   "_FPBmrA.comment": "Error message for source schema node not found",
   "_FSD5Z4.comment": "Aria label for select all workflows checkbox",
   "_FT3jt6.comment": "httpwebhook callback url section title",
