+++ libs/designer/src/lib/ui/panel/configureTemplatePanel/configureWorkflowsPanel/tabs/customizeWorkflowsTab.tsx
@@ -1,161 +1,11 @@
 import type { AppDispatch } from '../../../../../core/state/templates/store';
 import constants from '../../../../../common/constants';
-import { TemplatesSection, type TemplateTabProps, type TemplatesSectionItem } from '@microsoft/designer-ui';
+import type { TemplateTabProps } from '@microsoft/designer-ui';
 import { closePanel } from '../../../../../core/state/templates/panelSlice';
 import type { ConfigureWorkflowsTabProps } from '../configureWorkflowsPanel';
 import type { IntlShape } from 'react-intl';
 import type { WorkflowTemplateData } from '../../../../../core';
-import { Accordion, AccordionHeader, AccordionItem, AccordionPanel, Text } from '@fluentui/react-components';
-import { getResourceNameFromId, type Template } from '@microsoft/logic-apps-shared';
-import { useMemo } from 'react';
-import { useTemplatesStrings } from '../../../../templates/templatesStrings';
-
-export const CustomizeWorkflows = ({
-  selectedWorkflowsList,
-  updateWorkflowDataField,
-}: {
-  selectedWorkflowsList: Record<string, Partial<WorkflowTemplateData>>;
-  updateWorkflowDataField: (workflowId: string, workflowData: Partial<WorkflowTemplateData>) => void;
-}) => {
-  const workflowEntries = Object.entries(selectedWorkflowsList);
-
-  return (
-    <div className="msla-templates-tab msla-panel-no-description-tab">
-      {workflowEntries.length ? (
-        workflowEntries.length > 1 ? (
-          <Accordion multiple={true}>
-            {Object.entries(selectedWorkflowsList).map(([workflowId, workflowData]) => (
-              <AccordionItem value={workflowId} key={workflowId}>
-                <AccordionHeader>
-                  <Text style={{ fontWeight: 'bold' }}>{getResourceNameFromId(workflowId)}</Text>
-                </AccordionHeader>
-                <AccordionPanel>
-                  <CustomizeWorkflowSection
-                    workflowId={workflowId}
-                    isMultiWorkflowTemplate={true}
-                    workflow={workflowData}
-                    updateWorkflowDataField={updateWorkflowDataField}
-                  />
-                </AccordionPanel>
-              </AccordionItem>
-            ))}
-          </Accordion>
-        ) : (
-          <CustomizeWorkflowSection
-            workflowId={workflowEntries[0][0]}
-            isMultiWorkflowTemplate={false}
-            workflow={workflowEntries[0][1]}
-            updateWorkflowDataField={updateWorkflowDataField}
-          />
-        )
-      ) : null}
-    </div>
-  );
-};
-
-const CustomizeWorkflowSection = ({
-  workflowId,
-  isMultiWorkflowTemplate,
-  workflow,
-  updateWorkflowDataField,
-}: {
-  workflowId: string;
-  isMultiWorkflowTemplate: boolean;
-  workflow: Partial<WorkflowTemplateData>;
-  updateWorkflowDataField: (workflowId: string, workflowData: Partial<WorkflowTemplateData>) => void;
-}) => {
-  const { resourceStrings } = useTemplatesStrings();
-  const generalSectionItems: TemplatesSectionItem[] = useMemo(() => {
-    return [
-      {
-        label: resourceStrings.WORKFLOW_NAME,
-        value: workflow.workflowName || '',
-        type: 'textfield',
-        required: true,
-        onChange: (value: string) => {
-          updateWorkflowDataField(workflowId, { workflowName: value });
-        },
-      },
-      {
-        label: resourceStrings.WORKFLOW_DISPLAY_NAME,
-        value: workflow.manifest?.title || '',
-        type: 'textfield',
-        required: true,
-        onChange: (value: string) => {
-          updateWorkflowDataField(workflowId, {
-            ...workflow,
-            manifest: {
-              ...workflow.manifest,
-              title: value,
-            } as Template.WorkflowManifest,
-          });
-        },
-      },
-      //TODO: add state type
-    ];
-  }, [workflowId, updateWorkflowDataField, workflow, resourceStrings]);
-
-  const descriptionSectionItems: TemplatesSectionItem[] = useMemo(() => {
-    const baseItems: TemplatesSectionItem[] = isMultiWorkflowTemplate
-      ? [
-          {
-            label: resourceStrings.SUMMARY,
-            value: workflow.manifest?.summary || '',
-            type: 'textfield',
-            onChange: (value: string) => {
-              updateWorkflowDataField(workflowId, {
-                ...workflow,
-                manifest: {
-                  ...workflow.manifest,
-                  summary: value,
-                } as Template.WorkflowManifest,
-              });
-            },
-          },
-        ]
-      : [];
-    baseItems.push({
-      label: resourceStrings.DESCRIPTION,
-      value: workflow.manifest?.description || '',
-      type: 'textfield',
-      onChange: (value: string) => {
-        updateWorkflowDataField(workflowId, {
-          ...workflow,
-          manifest: {
-            ...workflow.manifest,
-            description: value,
-          } as Template.WorkflowManifest,
-        });
-      },
-    });
-    baseItems.push({
-      label: resourceStrings.PREREQUISITES,
-      value: workflow.manifest?.prerequisites || '',
-      type: 'textfield',
-      onChange: (value: string) => {
-        updateWorkflowDataField(workflowId, {
-          ...workflow,
-          manifest: {
-            ...workflow.manifest,
-            prerequisites: value,
-          } as Template.WorkflowManifest,
-        });
-      },
-    });
-    return baseItems;
-  }, [workflowId, updateWorkflowDataField, workflow, isMultiWorkflowTemplate, resourceStrings]);
-
-  return (
-    <div>
-      <TemplatesSection
-        title={isMultiWorkflowTemplate ? '' : resourceStrings.GENERAL}
-        titleHtmlFor={'generalSectionLabel'}
-        items={generalSectionItems}
-      />
-      <TemplatesSection title={resourceStrings.DESCRIPTION} titleHtmlFor={'descriptionSectionLabel'} items={descriptionSectionItems} />
-    </div>
-  );
-};
+import { CustomizeWorkflows } from '../../../../../ui/configuretemplate/workflows/customizeWorkflows';
 
 export const customizeWorkflowsTab = (
   intl: IntlShape,
@@ -167,7 +17,7 @@ export const customizeWorkflowsTab = (
     disabled,
     selectedWorkflowsList,
     updateWorkflowDataField,
-    onSave
+    onSave,
   }: ConfigureWorkflowsTabProps & {
     updateWorkflowDataField: (workflowId: string, workflowData: Partial<WorkflowTemplateData>) => void;
   }
